Add a  recently added view controller
